Get Account Balances Over Time ========================
Created the `MintHistoryReader` which reads the data stored in `/history`
and returns it in easily consumable data structures.
The only history it reads right now is the account balances over time
(a dictionary keyed by date retrieved where the value is a DataFrame of account data).

Added a `get_account_balances_over_time()` method to `MintAnalyzer`
that uses the `MintHistoryReader` to get the account balances over time
(a dictionary keyed by `(account_provider,account_name)` where the value is
a dictionary containing an object like `{'dates': [YYYY-MM-DD], 'balances': [$]}`
with a list of retrieval dates and a list of associated balances)

```json
{
    (MyBank,MyAccountName):
    {
        'dates': [YYYY-MM-DD,YYYY-MM-DD,YYYY-MM-DD],
        'balances': [$0.00,$0.00,$0.00]
    }
}
```